+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/cron/environment/ThirdPartyAgentHeartBeat.kt
@@ -24,7 +24,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.misc.service.environment
+package com.tencent.devops.misc.cron.environment
 
 import com.tencent.devops.common.api.enums.AgentAction
 import com.tencent.devops.common.api.enums.AgentStatus
